chat-server/requirements.txt
flask>=2.3.2
flask-socketio>=5.3.4
gunicorn>=20.1.0
gevent>=22.10.2
gevent-websocket>=0.10.1
grpcio>=1.54.2
langchain>=0.0.206
openai>=0.27.8
tiktoken>=0.4.0
redis>=4.5.5
